Updating to core 1685
