+++ src/NServiceBus.Hosting.Profiles/packages.config
@@ -2,7 +2,7 @@
 <packages>
   <package id="Fody" version="1.29.3" targetFramework="net452" developmentDependency="true" />
   <package id="GitVersionTask" version="2.0.1" targetFramework="net452" developmentDependency="true" />
-  <package id="NServiceBus" version="6.0.0-unstable1509" targetFramework="net452" />
+  <package id="NServiceBus" version="6.0.0-unstable1685" targetFramework="net452" />
   <package id="NuGet.CommandLine" version="2.8.3" targetFramework="net451" />
   <package id="NuGetPackager" version="0.5.5" targetFramework="net452" developmentDependency="true" />
   <package id="Obsolete.Fody" version="4.0.3" targetFramework="net452" developmentDependency="true" />
