+++ src/NServiceBus.Hosting.Tests/App.config
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <appSettings>
-    <add key="EndpointConfigurationType" value="NServiceBus.Hosting.Tests.ConfigWithCustomTransport, NServiceBus.Hosting.Tests"/>
+    <add key="EndpointConfigurationType" value="EndpointTypeDeterminerTests.MyEndpointConfig, NServiceBus.Hosting.Tests"/>
   </appSettings>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
