+++ src/NServiceBus.Hosting.Tests/MyEndpointConfig.cs
@@ -1,0 +1,12 @@
+namespace EndpointTypeDeterminerTests
+{
+    using NServiceBus;
+
+    class MyEndpointConfig : IConfigureThisEndpoint
+    {
+        public void Customize(EndpointConfiguration configuration)
+        {
+
+        }
+    }
+}
