+++ src/NServiceBus.Hosting.Tests/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NServiceBus" version="6.0.0-unstable1509" targetFramework="net452" />
+  <package id="NServiceBus" version="6.0.0-unstable1685" targetFramework="net452" />
   <package id="NUnit" version="2.6.4" targetFramework="net452" />
 </packages>
