+++ src/NServiceBus.Hosting.Windows/packages.config
@@ -4,7 +4,7 @@
   <package id="Fody" version="1.29.3" targetFramework="net452" developmentDependency="true" />
   <package id="GitVersionTask" version="2.0.1" targetFramework="net452" developmentDependency="true" />
   <package id="log4net" version="1.2.10" targetFramework="net45" />
-  <package id="NServiceBus" version="6.0.0-unstable1509" targetFramework="net452" />
+  <package id="NServiceBus" version="6.0.0-unstable1685" targetFramework="net452" />
   <package id="NuGet.CommandLine" version="2.8.3" targetFramework="net451" />
   <package id="NuGetPackager" version="0.5.5" targetFramework="net452" developmentDependency="true" />
   <package id="Obsolete.Fody" version="4.0.3" targetFramework="net452" developmentDependency="true" />
